Revert "Promote token_t to unsigned long long" This broke several things -- have to do this inside kenwood.c instead This reverts commit 6b3e749d332694b5fe95b0807744a33cc9950f07.

+++ include/hamlib/rig.h
@@ -832,7 +832,7 @@ typedef enum {
 /**
  * \brief configuration token
  */
-typedef unsigned long long token_t;
+typedef long token_t;
 
 
 //! @cond Doxygen_Suppress
